Refactored TerraVersion class.
